cliques/requirements.txt
appengine-toolkit==0.1.0
astroid==1.2.1
django-cors-headers==0.13
django-debug-toolbar==1.2.2
django-filter==0.8
Django==1.5.7
djangorestframework==2.4.4
flake8==2.2.5
logilab-common==0.63.0
mccabe==0.2.1
mock==1.0.1
MySQL-python==1.2.5
pep8==1.5.7
pip-tools==0.3.5
pusher==0.8
py==1.4.26
pyflakes==0.8.1
pylint-django==0.5.5
pylint-plugin-utils==0.2.2
pylint==1.3.1
requests==2.4.3
six==1.8.0
South==1.0.1
sqlparse==0.1.13
tox==1.8.1
urllib3==1.9.1
virtualenv==1.11.6
